refactored CLAMS workflow managing functions ...
- a new mmif.utils.workflow helper module to offer
    - common hashing implementation for app parameters
    - common CLAMS workflow ID calculation based on the hashing
    - summarizing MMIF file(s)
- CLI wrappers are provided as `mmif rewind` and `mmif describe` cmds
- all the CLI cmds are now properly show up in built documentation
