+++ documentation/autodoc/mmif.utils.rst
@@ -1,12 +1,15 @@
 mmif.utils package
 ==================
 
-Package containing utility modules for handling different types of source 
-documents, and general implementation of common data structures and 
+Package containing utility modules for handling different types of source
+documents, and general implementation of common data structures and
 algorithms.
 
+Submodules
+----------
+
 ``video_document_helper`` module
-----------------------------------------
+--------------------------------
 
 .. automodule:: mmif.utils.video_document_helper
    :members:
@@ -14,26 +17,33 @@ algorithms.
    :show-inheritance:
 
 ``text_document_helper`` module
----------------------------------
+-------------------------------
 
-.. automodule:: mmif.utils.sequence_helper
+.. automodule:: mmif.utils.text_document_helper
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+``timeunit_helper`` module
+-------------------------------
+
+.. automodule:: mmif.utils.timeunit_helper
    :members:
    :undoc-members:
    :show-inheritance:
 
 ``sequence_helper`` module
----------------------------------
+--------------------------
 
 .. automodule:: mmif.utils.sequence_helper
    :members:
    :undoc-members:
    :show-inheritance:
 
-``sequence_helper`` module
----------------------------------
+``workflow_helper`` module
+--------------------------
 
-.. automodule:: mmif.utils.sequence_helper
+.. automodule:: mmif.utils.workflow_helper
    :members:
    :undoc-members:
    :show-inheritance:
-
