+++ mmif/utils/cli/describe.py
@@ -1,249 +1,19 @@
 import argparse
-import hashlib
 import json
 import sys
 import textwrap
-import datetime
-from collections import Counter, defaultdict
 from pathlib import Path
-from typing import Union, List, Tuple, Optional, Any
+from typing import Union
 
-from mmif import Mmif
-
-
-def _split_appname_appversion(
-        long_app_id: str
-) -> Tuple[Optional[str], Optional[str]]:
-    """
-    Split app name and version from a long app identifier.
-
-    Assumes the identifier looks like "uri://APP_DOMAIN/APP_NAME/APP_VERSION"
-
-    :param long_app_id: Full app identifier URI
-    :return: Tuple of (app_name, app_version), either may be None if not found
-    """
-    app_path = Path(long_app_id).parts
-    app_name = app_path[2] if len(app_path) > 2 else None
-    app_version = app_path[3] if len(app_path) > 3 else None
-    if (app_version is not None and app_name is not None
-            and app_name.endswith(app_version)):
-        app_name = app_name[:-len(app_version) - 1]
-    if app_version == 'unresolvable':
-        app_version = None
-    return app_name, app_version
-
-
-def _get_profile_data(view) -> dict:
-    """
-    Extract profiling data from a view's metadata.
-
-    :param view: MMIF view object
-    :return: Dictionary of profiling data
-    """
-    # TODO (krim @ 2025-11-27): the GPU part is heavily rely on how clams-python implements _cuda_memory_to_str funct
-    # also it's not clear how helpful vram usage in the describe output is
-    # So I'm not using vram records here. Perhaps should `describe` be moved to clams-python instead?
-
-    # running time can be found two ways: either in appProfiling.runningTime or appRunningTime (legacy) key
-    profiling = view.metadata.get("appProfiling", {})
-    if "runningTime" not in profiling:
-        running_time_str = view.metadata.get("appRunningTime")
-    else:
-        running_time_str = profiling.get("runningTime")
-        
-    if running_time_str is None:
-        return {}
-
-    # the format is datetime.timedelta string, e.g. '0:00:02.345678'
-    # need to convert to milliseconds integer
-    time_obj = datetime.datetime.strptime(running_time_str, "%H:%M:%S.%f").time()
-    milliseconds = (time_obj.hour * 3600 + time_obj.minute * 60 + time_obj.second) * 1000 + time_obj.microsecond // 1000
-    return {"runningTime": milliseconds}
-
-
-def _group_views_by_app_execution(views: List[Any]) -> List[List[Any]]:
-    """
-    Groups views into app executions based on app and timestamp.
-
-    An "app execution" is a set of views produced by the same app at the
-    exact same timestamp.
-    """
-    import itertools
-    # Filter out views that don't have a timestamp or app, as they can't be grouped.
-    groupable_views = [
-        v for v in views 
-        if v.metadata.get("app") and v.metadata.get("timestamp") is not None
-    ]
-    
-    # Sort views by timestamp first, then by app URI to ensure deterministic grouping
-    groupable_views.sort(key=lambda v: (v.metadata.timestamp, v.metadata.app))
-
-    # Group by app and timestamp
-    app_executions = []
-    for key, group in itertools.groupby(groupable_views, key=lambda v: (v.metadata.app, v.metadata.timestamp)):
-        app_executions.append(list(group))
-        
-    return app_executions
+from mmif.utils.workflow_helper import generate_workflow_identifier, describe_single_mmif, \
+    describe_mmif_collection
 
 
 def get_pipeline_specs(mmif_file: Union[str, Path]):
     import warnings
-    warnings.warn("get_pipeline_specs is deprecated, use describe_single_mmif instead", DeprecationWarning)
+    warnings.warn("get_pipeline_specs is deprecated, use mmif.utils.workflow_helper.describe_single_mmif instead",
+                  DeprecationWarning)
     return describe_single_mmif(mmif_file)
-
-
-def describe_single_mmif(mmif_file: Union[str, Path]) -> dict:
-    """
-    Reads a MMIF file and extracts the workflow specification from it.
-
-    This function provides an app-centric summarization of the workflow. The
-    conceptual hierarchy is that a **workflow** is a sequence of **apps**,
-    and each **app** execution can produce one or more **views**. This function
-    groups views that share the same ``app`` and ``metadata.timestamp`` into
-    a single logical "app execution".
-
-    .. note::
-        For MMIF files generated by ``clams-python`` <= 1.3.3, all views
-        are independently timestamped. This means that even if multiple views
-        were generated by a single execution of an app, their
-        ``metadata.timestamp`` values will be unique. As a result, the grouping
-        logic will treat each view as a separate app execution. The change
-        that aligns timestamps for views from a single app execution is
-        implemented in `clams-python PR #271
-        <https://github.com/clamsproject/clams-python/pull/271>`_.
-
-    The output format is a dictionary with the following keys:
-
-    * **workflowId**: A unique identifier for the workflow, based on the
-      sequence of app executions (app, version, parameter hashes). App
-      executions with errors or warnings are excluded from this identifier.
-    * **stats**:
-        * **appCount**: Total number of identified app executions.
-        * **errorViews**: A list of view IDs that reported errors.
-        * **warningViews**: A list of view IDs that reported warnings.
-        * **emptyViews**: A list of view IDs that contain no annotations.
-        * **annotationCount**: A dictionary with the ``total`` number of
-          annotations across all app executions.
-        * **annotationCountByType**: A dictionary mapping each annotation
-          type to its ``total`` count across all app executions.
-    * **apps**: A list of objects, where each object represents one app
-      execution. It includes metadata, profiling, and aggregated statistics
-      for all views generated by that execution. A special entry for views
-      that could not be assigned to an execution will be at the end of the list.
-
-    :param mmif_file: Path to the MMIF file
-    :return: A dictionary containing the workflow specification.
-    """
-    if not isinstance(mmif_file, (str, Path)):
-        raise ValueError(
-            "MMIF file path must be a string or a Path object."
-        )
-
-    workflow_id = generate_workflow_identifier(mmif_file)
-    with open(mmif_file, "r") as f:
-        mmif_str = f.read()
-
-    mmif = Mmif(mmif_str)
-
-    error_view_ids = []
-    warning_view_ids = []
-    empty_view_ids = []
-    
-    # Generate the new "apps" list
-    app_executions = []
-    processed_view_ids = set()
-    view_groups = _group_views_by_app_execution(mmif.views)
-    for group in view_groups:
-        first_view = group[0]
-        # skip executions with errors or warnings
-        if first_view.has_error() or first_view.has_warnings():
-            continue
-            
-        execution_ann_counter = Counter()
-        for view in group:
-            if len(view.annotations) == 0:
-                empty_view_ids.append(view.id)
-            execution_ann_counter.update(Counter(str(ann.at_type) for ann in view.annotations))
-        
-        execution_ann_by_type = {at_type: {"total": count} for at_type, count in execution_ann_counter.items()}
-        
-        execution_view_ids = [v.id for v in group]
-        processed_view_ids.update(execution_view_ids)
-
-        app_executions.append({
-            "app": first_view.metadata.app,
-            "viewIds": execution_view_ids,
-            "appConfiguration": first_view.metadata.get("appConfiguration", {}),
-            "appProfiling": _get_profile_data(first_view)
-        })
-        
-    # Handle unassigned and problematic views
-    all_view_ids = set(v.id for v in mmif.views)
-        
-    for view in mmif.views:
-        if view.id not in processed_view_ids:
-            if view.has_error():
-                error_view_ids.append(view.id)
-            elif view.has_warnings():
-                warning_view_ids.append(view.id)
-            elif len(view.annotations) == 0:
-                empty_view_ids.append(view.id)
-
-    unassigned_view_ids = all_view_ids - processed_view_ids - set(error_view_ids) - set(warning_view_ids)
-    
-    # Store app_execution_count before potentially adding the special entry
-    app_execution_count = len(app_executions) 
-
-    if unassigned_view_ids:
-        app_executions.append({
-            "app": "http://apps.clams.ai/non-existing-app/v1",
-            "viewIds": sorted(list(unassigned_view_ids))
-        })
-
-    # aggregate total annotation counts
-    total_annotations = 0
-    total_annotations_by_type = defaultdict(lambda: {'total': 0})
-    for execution in app_executions:
-        # Only aggregate from actual app executions, not the special unassigned entry
-        if execution.get('app') != "http://apps.clams.ai/non-existing-app/v1":
-            if 'annotationCount' in execution:
-                total_annotations += execution['annotationCount']['total']
-            if 'annotationCountByType' in execution:
-                for at_type, data in execution['annotationCountByType'].items():
-                    total_annotations_by_type[at_type]['total'] += data['total']
-
-    return {
-        "workflowId": workflow_id,
-        "stats": {
-            "appCount": app_execution_count,
-            "errorViews": error_view_ids,
-            "warningViews": warning_view_ids,
-            "emptyViews": empty_view_ids,
-            "annotationCount": {"total": total_annotations},
-            "annotationCountByType": dict(total_annotations_by_type)
-        },
-        "apps": app_executions
-    }
-
-
-def generate_param_hash(params: dict) -> str:
-    """
-    Generate MD5 hash from a parameter dictionary.
-
-    Parameters are sorted alphabetically, joined as key=value pairs,
-    and hashed using MD5. This is not for security purposes, only for
-    generating consistent identifiers.
-
-    :param params: Dictionary of parameters
-    :return: MD5 hash string (32 hex characters)
-    """
-    if not params:
-        param_string = ""
-    else:
-        param_list = ['='.join([k, str(v)]) for k, v in params.items()]
-        param_list.sort()
-        param_string = ','.join(param_list)
-    return hashlib.md5(param_string.encode('utf-8')).hexdigest()
 
 
 def generate_pipeline_identifier(mmif_file: Union[str, Path]) -> str:
@@ -251,64 +21,6 @@ def generate_pipeline_identifier(mmif_file: Union[str, Path]) -> str:
     warnings.warn("generate_pipeline_identifier is deprecated, use generate_workflow_identifier instead",
                   DeprecationWarning)
     return generate_workflow_identifier(mmif_file)
-
-
-def generate_workflow_identifier(mmif_file: Union[str, Path]) -> str:
-    """
-    Generate a workflow identifier string from a MMIF file.
-
-    The identifier follows the storage directory structure format:
-    app_name/version/param_hash/app_name2/version2/param_hash2/...
-
-    Uses view.metadata.parameters (raw user-passed values) for hashing
-    to ensure reproducibility. Views with errors or warnings are excluded
-    from the identifier; empty views are included.
-    """
-    if not isinstance(mmif_file, (str, Path)):
-        raise ValueError(
-            "MMIF file path must be a string or a Path object."
-        )
-
-    with open(mmif_file, "r") as f:
-        mmif_str = f.read()
-
-    data = Mmif(mmif_str)
-    segments = []
-
-    # First prefix is source information, sorted by document type
-    sources = Counter(doc.at_type.shortname for doc in data.documents)
-    segments.append('-'.join([f'{k}-{sources[k]}' for k in sorted(sources.keys())]))
-
-    # Group views into runs
-    app_executions = _group_views_by_app_execution(data.views)
-    
-    for app_execution in app_executions:
-        # Use the first view in the run as representative for metadata
-        first_view = app_execution[0]
-
-        # Skip runs where the representative view has errors or warnings
-        if first_view.has_error() or first_view.has_warnings():
-            continue
-
-        app = first_view.metadata.get("app")
-        if app is None:
-            continue
-        app_name, app_version = _split_appname_appversion(app)
-
-        # Use raw parameters from the first view for reproducibility
-        try:
-            param_dict = first_view.metadata.parameters
-        except (KeyError, AttributeError):
-            param_dict = {}
-
-        param_hash = generate_param_hash(param_dict)
-
-        # Build segment: app_name/version/hash
-        name_str = app_name if app_name else "unknown"
-        version_str = app_version if app_version else "unversioned"
-        segments.append(f"{name_str}/{version_str}/{param_hash}")
-
-    return '/'.join(segments)
 
 
 def describe_argparser():
@@ -321,26 +33,28 @@ def describe_argparser():
         'provides CLI to describe the workflow specification from a MMIF '
         'file or a collection of MMIF files.'
     )
-    additional = textwrap.dedent("""
-    MMIF describe extracts workflow information from a single MMIF file or 
+
+    # get and clean docstrings
+    single_doc = describe_single_mmif.__doc__.split(':param')[0]
+    single_doc = textwrap.dedent(single_doc).strip()
+    collection_doc = describe_mmif_collection.__doc__.split(':param')[0]
+    collection_doc = textwrap.dedent(collection_doc).strip()
+
+    additional = textwrap.dedent(f"""
+    This command extracts workflow information from a single MMIF file or 
     summarizes a directory of MMIF files.
+    
+    ==========================
+    For a single MMIF file
+    ==========================
+    {single_doc}
 
-    For a single MMIF file, the output is a JSON summary including:
-    - workflowId: unique identifier for the workflow based on source document, 
-      apps, versions, and parameter hashes (excludes app executions with errors/warnings)
-    - stats: summary statistics about app executions, error/warning/empty views, and total annotation counts
-    - apps: a list of app execution objects, each detailing app metadata, profiling, and view IDs
-
-    For a directory, the output is a JSON summary of all MMIF files, including:
-    - mmifCountByStatus: counts of total, successful, with errors, with warnings, and invalid MMIF files
-    - mmifCountByWorkflow: counts of MMIFs grouped by their unique workflowId
-    - appProfilings: aggregated profiling statistics (appCount, avgRunningTimeMs, etc.) for each app URI across the collection
-    - annotationCountByType: aggregated total annotation counts per type across the collection
-
-    App executions with errors or warnings are tracked but excluded from the workflow
-    identifier and overall app execution statistics."""
-                                 )
-    return oneliner, oneliner + '\n\n' + additional
+    ===============================
+    For a directory of MMIF files
+    ===============================
+    {collection_doc}
+    """)
+    return oneliner, oneliner + '\n\n' + additional.strip()
 
 
 def prep_argparser(**kwargs):
@@ -368,81 +82,6 @@ def prep_argparser(**kwargs):
         help="Pretty-print JSON output"
     )
     return parser
-
-
-def describe_mmif_collection(mmif_dir: Union[str, Path]) -> dict:
-    """
-    Read all MMIF files in a directory and extract the workflow specifications.
-    """
-    import statistics
-    import itertools
-    from collections import defaultdict
-
-    mmif_files = Path(mmif_dir).glob('*.mmif')
-
-    status_summary = defaultdict(int)
-    status_summary['total'] = 0
-    status_summary['successful'] = 0
-    status_summary['withErrors'] = 0
-    status_summary['withWarnings'] = 0
-    status_summary['invalid'] = 0
-
-    workflow_analysis = defaultdict(lambda: {'count': 0})
-    app_profilings_raw = defaultdict(list)
-    annotation_counts = defaultdict(int)
-    annotation_counts['total'] = 0
-
-    for mmif_file in mmif_files:
-        status_summary['total'] += 1
-        try:
-            single_report = describe_single_mmif(mmif_file)
-        except Exception as e:
-            status_summary['invalid'] += 1
-            continue
-
-        if single_report['stats']['errorViews']:
-            status_summary['withErrors'] += 1
-        elif single_report['stats']['warningViews']:
-            status_summary['withWarnings'] += 1
-        else:
-            status_summary['successful'] += 1
-
-        # Workflow analysis
-        wf_id = single_report['workflowId']
-        workflow_analysis[wf_id]['count'] += 1
-
-        # App performance and annotation raw data collection from "apps" list
-        for app_execution in single_report.get('apps', []):
-            # app profilings
-            app = app_execution.get('app')
-            profiling = app_execution.get('appProfiling', {})
-            running_time = profiling.get('runningTime')
-            if app and running_time is not None:
-                app_profilings_raw[app].append(running_time)
-            
-            # annotation counts
-            annotation_counts['total'] += app_execution.get('annotationCount', {}).get('total', 0)
-            for at_type, data in app_execution.get('annotationCountByType', {}).items():
-                annotation_counts[at_type] += data.get('total', 0)
-
-    # Process app performance data
-    profiles = {}
-    for app, execution_times in app_profilings_raw.items():
-        if execution_times:
-            profiles[app] = {
-                'appCount': len(execution_times),
-                'avgRunningTimeMs': statistics.mean(execution_times),
-                'minRunningTimeMs': min(execution_times),
-                'maxRunningTimeMs': max(execution_times),
-                'stdevRunningTimeMs': statistics.stdev(execution_times) if len(execution_times) > 1 else 0
-            }
-
-    return {
-        'mmifCountByStatus': dict(status_summary),
-        'mmifCountByWorkflow': {k: v for k, v in sorted(workflow_analysis.items())},
-        'appProfilings': profiles,
-        'annotationCountByType': dict(annotation_counts)
-    }
 
 
 def main(args):
